Renamed test files and test classes per standard.
