+++ tests/Integration/CloudflareSubscriber/setVarnishLocalhost.php
