+++ tests/Unit/Cloudflare/hasPageRule.php
@@ -8,15 +8,15 @@ use WPMedia\Cloudflare\CloudflareFacade;
 use WPMedia\Cloudflare\Tests\Unit\TestCase;
 
 /**
- * @covers WPMedia\Cloudflare\Cloudflare::get_cloudflare_ips
+ * @covers WPMedia\Cloudflare\Cloudflare::has_page_rule
  * @group  Addon
  */
-class TestGetCloudflareIPS extends TestCase {
+class Test_HasPageRule extends TestCase {
 
 	/**
-	 * Test get cloudflare IPs with cached invalid transient for credentials.
+	 * Test Cloudflare has page rules with cached invalid transient.
 	 */
-	public function testGetCloudflareIPSWithInvalidCredentials() {
+	public function testHasRuleWithInvalidCredentials() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -30,55 +30,24 @@ class TestGetCloudflareIPS extends TestCase {
 		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-
-		Functions\when( 'get_transient' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
-		$cloudflare_facade_mock->shouldReceive('ips')->andThrow( new \Exception() );
-		Functions\expect( 'set_transient' )->once();
+		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
 
 		$this->assertEquals(
-			$mocks[ 'cf_ips' ],
-			$cloudflare->get_cloudflare_ips()
+			$wp_error,
+			$has_page_rule
 		);
 	}
 
 	/**
-	 * Test get cloudflare IPs with invalid credentials and cached IPs in transient `rocket_cloudflare_ips`.
+	 * Test Cloudflare has page rules with exception.
 	 */
-	public function testGetCloudflareIPSWithInvalidCredentialsButIPSCached() {
+	public function testHasRuleWithException() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
 		$wp_error               = $mocks['wp_error'];
 
 		// The Cloudflare constructor run with transient set as WP_Error.
-		Functions\when( 'get_transient' )->justReturn( $wp_error );
-		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
-		Functions\expect( 'set_transient' )->never();
-		Functions\when( 'is_wp_error' )->justReturn( true );
-		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
-
-		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-
-		Functions\when( 'get_transient' )->justReturn( $mocks[ 'cf_ips' ] );
-		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
-
-		$this->assertEquals(
-			$mocks[ 'cf_ips' ],
-			$cloudflare->get_cloudflare_ips()
-		);
-	}
-
-	/**
-	 * The get Cloudflare IPs with valid CF credentials, no cached `rocket_cloudflare_ips` and error on `ips()`.
-	 */
-	public function testGetCloudflareIPSWithValidCredentialsAndNoCachedIPSWithError() {
-		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
-
-		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error               = $mocks['wp_error'];
-
-		 // The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
 		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
 		Functions\expect( 'set_transient' )->never();
@@ -86,29 +55,83 @@ class TestGetCloudflareIPS extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-
-		Functions\when( 'get_transient' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
-		$cf_reply = json_decode('{"success":false,"errors":[{"code":1007,"message":"Invalid value"}],"messages":[],"result":null}');
-		$cloudflare_facade_mock->shouldReceive('ips')->andReturn( $cf_reply );
-		Functions\expect( 'set_transient' )->once();
+		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andThrow( new \Exception() );
+		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
 
 		$this->assertEquals(
-			$mocks[ 'cf_ips' ],
-			$cloudflare->get_cloudflare_ips()
+			new \WP_Error(),
+			$has_page_rule
 		);
 	}
 
+
 	/**
-	 * The get Cloudflare IPs with valid CF credentials, no cached `rocket_cloudflare_ips` and success `ips()`.
+	 * Test Cloudflare has page rules with no success.
 	 */
-	public function testGetCloudflareIPSWithValidCredentialsAndNoCachedIPSWithSuccess() {
+	public function testHasRuleWithNoSuccess() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
 		$wp_error               = $mocks['wp_error'];
 
-		 // The Cloudflare constructor run with transient set as WP_Error.
+		// The Cloudflare constructor run with transient set as WP_Error.
+		Functions\when( 'get_transient' )->justReturn( true );
+		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
+		Functions\expect( 'set_transient' )->never();
+		Functions\when( 'is_wp_error' )->justReturn( false );
+		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
+
+		Functions\when( 'wp_sprintf_l' )->justReturn( '' );
+		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+		$cf_page_rule = json_decode('{"success":false,"errors":[{"code":7003,"message":"Could not route to \/zones\/ZONE_ID, perhaps your object identifier is invalid?"},{"code":7000,"message":"No route for that URI"}],"messages":[],"result":null}');
+		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
+		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
+
+		$this->assertEquals(
+			new \WP_Error(),
+			$has_page_rule
+		);
+	}
+
+	/**
+	 * Test Cloudflare has page rules with success but no page rule.
+	 */
+	public function testHasRuleWithSuccessButNoPageRule() {
+		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
+
+		$cloudflare_facade_mock = $mocks['facade'];
+		$wp_error               = $mocks['wp_error'];
+
+		// The Cloudflare constructor run with transient set as WP_Error.
+		Functions\when( 'get_transient' )->justReturn( true );
+		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
+		Functions\expect( 'set_transient' )->never();
+		Functions\when( 'is_wp_error' )->justReturn( false );
+
+		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
+
+		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+		$cf_page_rule = json_decode('{"result":[{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"bypass"}],"priority":3,"status":"active","created_on":"","modified_on":""},{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":""}],"priority":2,"status":"active","created_on":"","modified_on":""}],"success":true,"errors":[],"messages":[]}');
+		Functions\when( 'wp_json_encode' )->justReturn( json_encode( $cf_page_rule ) );
+		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
+		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
+
+		$this->assertEquals(
+			false,
+			$has_page_rule
+		);
+	}
+
+	/**
+	 * Test Cloudflare has page rules with success and page rule.
+	 */
+	public function testHasRuleWithSuccessAndPageRule() {
+		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
+
+		$cloudflare_facade_mock = $mocks['facade'];
+		$wp_error               = $mocks['wp_error'];
+
+		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
 		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
 		Functions\expect( 'set_transient' )->never();
@@ -116,22 +139,14 @@ class TestGetCloudflareIPS extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-
-		Functions\when( 'get_transient' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
-		$cf_reply = json_decode('{"result":{"ipv4_cidrs":["173.245.48.0/20","103.21.244.0/22","103.22.200.0/22","103.31.4.0/22","141.101.64.0/18","108.162.192.0/18","190.93.240.0/20","188.114.96.0/20","197.234.240.0/22","198.41.128.0/17","162.158.0.0/15","104.16.0.0/12","172.64.0.0/13","131.0.72.0/22"],"ipv6_cidrs":["2400:cb00::/32","2606:4700::/32","2803:f800::/32","2405:b500::/32","2405:8100::/32","2a06:98c0::/29","2c0f:f248::/32"],"etag":"fb21705459fea38d23b210ee7d67b753"},"success":true,"errors":[],"messages":[]}');
-		$cloudflare_facade_mock->shouldReceive('ips')->andReturn( $cf_reply );
-		Functions\expect( 'set_transient' )->once();
-
-		$ips = $cloudflare->get_cloudflare_ips();
+		$cf_page_rule = json_decode('{"result":[{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"bypass"}],"priority":3,"status":"active","created_on":"","modified_on":""},{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"cache_everything"}],"priority":2,"status":"active","created_on":"","modified_on":""}],"success":true,"errors":[],"messages":[]}');
+		Functions\when( 'wp_json_encode' )->justReturn( json_encode( $cf_page_rule ) );
+		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
+		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
 
 		$this->assertEquals(
-			$mocks[ 'cf_ips' ]->result->ipv4_cidrs,
-			$ips->result->ipv4_cidrs
-		);
-		$this->assertEquals(
-			$mocks[ 'cf_ips' ]->result->ipv6_cidrs,
-			$ips->result->ipv6_cidrs
+			true,
+			$has_page_rule
 		);
 	}
 
@@ -177,45 +192,10 @@ class TestGetCloudflareIPS extends TestCase {
 		$facade   = Mockery::mock( CloudflareFacade::class );
 		$wp_error = Mockery::mock( 'WP_Error' );
 
-		$cf_ips = (object) [
-			'result'   => (object) [],
-			'success'  => true,
-			'errors'   => [],
-			'messages' => [],
-		];
-
-		$cf_ips->result->ipv4_cidrs = [
-			'173.245.48.0/20',
-			'103.21.244.0/22',
-			'103.22.200.0/22',
-			'103.31.4.0/22',
-			'141.101.64.0/18',
-			'108.162.192.0/18',
-			'190.93.240.0/20',
-			'188.114.96.0/20',
-			'197.234.240.0/22',
-			'198.41.128.0/17',
-			'162.158.0.0/15',
-			'104.16.0.0/12',
-			'172.64.0.0/13',
-			'131.0.72.0/22',
-		];
-
-		$cf_ips->result->ipv6_cidrs = [
-			'2400:cb00::/32',
-			'2606:4700::/32',
-			'2803:f800::/32',
-			'2405:b500::/32',
-			'2405:8100::/32',
-			'2a06:98c0::/29',
-			'2c0f:f248::/32',
-		];
-
 		$mocks = [
 			'options'  => $options,
 			'facade'   => $facade,
 			'wp_error' => $wp_error,
-			'cf_ips'   => $cf_ips,
 		];
 		return $mocks;
 	}
