+++ tests/Unit/Cloudflare/purgeByUrl.php
@@ -8,15 +8,15 @@ use WPMedia\Cloudflare\CloudflareFacade;
 use WPMedia\Cloudflare\Tests\Unit\TestCase;
 
 /**
- * @covers WPMedia\Cloudflare\Cloudflare::set_devmode
+ * @covers WPMedia\Cloudflare\Cloudflare::purge_by_url
  * @group  Addon
  */
-class TestSetDevMode extends TestCase {
+class Test_PurgeByUrl extends TestCase {
 
 	/**
-	 * Test set dev mode with cached invalid transient.
+	 * Test purge by url Cloudflare with cached invalid transient.
 	 */
-	public function testSetDevModeWithInvalidCredentials() {
+	public function testPurgeCloudflareByUrlWithInvalidCredentials() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -33,14 +33,14 @@ class TestSetDevMode extends TestCase {
 
 		$this->assertEquals(
 			$wp_error,
-			$cloudflare->set_devmode( false )
+			$cloudflare->purge_by_url( null, [ '/purge-url' ], null )
 		);
 	}
 
 	/**
-	 * Test set dev mode with exception.
+	 * Test purge by url Cloudflare with exception.
 	 */
-	public function testSetDevModeWithException() {
+	public function testPurgeCloudflareByUrlWithException() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -53,19 +53,19 @@ class TestSetDevMode extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andThrow( new \Exception() );
+		$cloudflare_facade_mock->shouldReceive('purge_files')->andThrow( new \Exception() );
 
 		$this->assertEquals(
 			new \WP_Error(),
-			$cloudflare->set_devmode( false )
+			$cloudflare->purge_by_url( null, [ '/purge-url' ], null )
 		);
 	}
 
 
 	/**
-	 * Test set dev mode with no success.
+	 * Test purge by url Cloudflare with no success.
 	 */
-	public function testSetDevModeWithNoSuccess() {
+	public function testPurgeCloudflareByUrlWithNoSuccess() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -79,19 +79,19 @@ class TestSetDevMode extends TestCase {
 
 		Functions\when( 'wp_sprintf_l' )->justReturn( '' );
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cf_reply   = json_decode('{"success":false,"errors":[{"code":1007,"message":"Invalid value for zone setting development_mode"}],"messages":[],"result":null}');
-		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andReturn( $cf_reply );
+		$cf_purge   = json_decode('{"success":false,"errors":[{"code":7001,"message":"Method GET not available for that URI."}],"messages":[],"result":null}');
+		$cloudflare_facade_mock->shouldReceive('purge_files')->andReturn( $cf_purge );
 
 		$this->assertEquals(
 			new \WP_Error(),
-			$cloudflare->set_devmode( false )
+			$cloudflare->purge_by_url( null, [ '/purge-url' ], null )
 		);
 	}
 
 	/**
-	 * Test set dev mode with success.
+	 * Test purge by url Cloudflare with success.
 	 */
-	public function testSetDevModeWithSuccess() {
+	public function testPurgeCloudflareByUrlWithSuccess() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -104,12 +104,12 @@ class TestSetDevMode extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cf_reply = json_decode('{"result":{"id":"development_mode","value":"off","modified_on":"","time_remaining":0,"editable":true},"success":true,"errors":[],"messages":[]}');
-		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andReturn( $cf_reply );
+		$cf_purge = json_decode('{"success": true,"errors": [],"messages": [],"result": {"id": ""}}');
+		$cloudflare_facade_mock->shouldReceive('purge_files')->andReturn( $cf_purge );
 
 		$this->assertEquals(
-			'off',
-			$cloudflare->set_devmode( false )
+			true,
+			$cloudflare->purge_by_url( null, [ '/purge-url' ], null )
 		);
 	}
