+++ tests/Unit/Cloudflare/setDevMode.php
@@ -8,56 +8,21 @@ use WPMedia\Cloudflare\CloudflareFacade;
 use WPMedia\Cloudflare\Tests\Unit\TestCase;
 
 /**
- * @covers WPMedia\Cloudflare\Cloudflare::get_instance
+ * @covers WPMedia\Cloudflare\Cloudflare::set_devmode
  * @group  Addon
  */
-class TestGetInstance extends TestCase {
+class Test_SetDevMode extends TestCase {
 
 	/**
-	 * Test Get Instance with Invalid Cloudflare credentials and no transient set.
+	 * Test set dev mode with cached invalid transient.
 	 */
-	public function testGetInstanceWithInvalidCFCredentialsNoTransient() {
+	public function testSetDevModeWithInvalidCredentials() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error   = \Mockery::mock( \WP_Error::class );
+		$wp_error               = $mocks['wp_error'];
 
-		Functions\when( 'get_transient' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('is_api_keys_valid')->andReturn( $wp_error );
-		Functions\expect( 'set_transient' )->once();
-		Functions\when( 'is_wp_error' )->justReturn( true );
-		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
-
-		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-	}
-
-	/**
-	 * Test Get Instance with valid Cloudflare credentials and no transient set.
-	 */
-	public function testGetInstanceWithValidCFCredentialsNoTransient() {
-		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
-
-		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error   = \Mockery::mock( \WP_Error::class );
-
-		Functions\when( 'get_transient' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('is_api_keys_valid')->andReturn( true );
-		Functions\expect( 'set_transient' )->once();
-		Functions\when( 'is_wp_error' )->justReturn( false );
-		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
-
-		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-	}
-
-	/**
-	 * Test Get Instance with invalid Cloudflare credentials and transient set.
-	 */
-	public function testGetInstanceWithInValidCFCredentialsAndTransient() {
-		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
-
-		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error   = \Mockery::mock( \WP_Error::class );
-
+		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( $wp_error );
 		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
 		Functions\expect( 'set_transient' )->never();
@@ -65,24 +30,88 @@ class TestGetInstance extends TestCase {
 		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+
+		$this->assertEquals(
+			$wp_error,
+			$cloudflare->set_devmode( false )
+		);
 	}
 
 	/**
-	 * Test Get Instance with valid Cloudflare credentials and transient set.
+	 * Test set dev mode with exception.
 	 */
-	public function testGetInstanceWithValidCFCredentialsAndTransient() {
+	public function testSetDevModeWithException() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
+
+		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
 		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
 		Functions\expect( 'set_transient' )->never();
-		Functions\when( 'is_wp_error' )->justReturn( true );
-		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
+		Functions\when( 'is_wp_error' )->justReturn( false );
+		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andThrow( new \Exception() );
+
+		$this->assertEquals(
+			new \WP_Error(),
+			$cloudflare->set_devmode( false )
+		);
 	}
 
+
+	/**
+	 * Test set dev mode with no success.
+	 */
+	public function testSetDevModeWithNoSuccess() {
+		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
+
+		$cloudflare_facade_mock = $mocks['facade'];
+
+		// The Cloudflare constructor run with transient set as WP_Error.
+		Functions\when( 'get_transient' )->justReturn( true );
+		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
+		Functions\expect( 'set_transient' )->never();
+		Functions\when( 'is_wp_error' )->justReturn( false );
+		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
+
+		Functions\when( 'wp_sprintf_l' )->justReturn( '' );
+		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+		$cf_reply   = json_decode('{"success":false,"errors":[{"code":1007,"message":"Invalid value for zone setting development_mode"}],"messages":[],"result":null}');
+		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andReturn( $cf_reply );
+
+		$this->assertEquals(
+			new \WP_Error(),
+			$cloudflare->set_devmode( false )
+		);
+	}
+
+	/**
+	 * Test set dev mode with success.
+	 */
+	public function testSetDevModeWithSuccess() {
+		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
+
+		$cloudflare_facade_mock = $mocks['facade'];
+
+		// The Cloudflare constructor run with transient set as WP_Error.
+		Functions\when( 'get_transient' )->justReturn( true );
+		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
+		Functions\expect( 'set_transient' )->never();
+		Functions\when( 'is_wp_error' )->justReturn( false );
+		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
+
+		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
+		$cf_reply = json_decode('{"result":{"id":"development_mode","value":"off","modified_on":"","time_remaining":0,"editable":true},"success":true,"errors":[],"messages":[]}');
+		$cloudflare_facade_mock->shouldReceive('change_development_mode')->andReturn( $cf_reply );
+
+		$this->assertEquals(
+			'off',
+			$cloudflare->set_devmode( false )
+		);
+	}
 
 	/**
 	 * Get the mocks required by Cloudflare’s constructor.
@@ -124,10 +153,12 @@ class TestGetInstance extends TestCase {
 		$options->method('get')->will( $this->returnValueMap( $map ) );
 
 		$facade   = Mockery::mock( CloudflareFacade::class );
+		$wp_error = Mockery::mock( 'WP_Error' );
 
 		$mocks = [
-			'options' => $options,
-			'facade'  => $facade,
+			'options'  => $options,
+			'facade'   => $facade,
+			'wp_error' => $wp_error,
 		];
 		return $mocks;
 	}
